requirements.txt
amqp==5.2.0
aniso8601==9.0.1
Babel==2.13.1
bcrypt==4.0.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.5
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.3
MarkupSafe==2.1.3
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
pycparser==2.21
pydyf==0.8.0
pyparsing==3.1.1
pyphen==0.14.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
setuptools==68.2.2
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.23
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.10
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
zopfli==0.2.3
